requirements.txt
alembic==1.7.4
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
coverage==6.0.2
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
flask-talisman==0.8.1
Flask-UUID==0.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow==3.14.0
marshmallow-sqlalchemy==0.26.1
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-cov==3.0.0
pytest-flask==1.2.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
toml==0.10.2
tomli==1.2.1
urllib3==1.26.7
Werkzeug==2.0.2
